+++ services/atlas-consumables/atlas.com/consumables/monster/processor.go
@@ -20,7 +20,7 @@ func NewProcessor(l logrus.FieldLogger, ctx context.Context) *Processor {
 	return p
 }
 
-func (p *Processor) CreateMonster(f field.Model, monsterId uint32, x int16, y int16, fh uint16, team int32) error {
+func (p *Processor) CreateMonster(f field.Model, monsterId uint32, x int16, y int16, fh int16, team int8) error {
 	_, err := requestCreate(f, monsterId, x, y, fh, team)(p.l, p.ctx)
 	if err != nil {
 		p.l.WithError(err).Errorf("Creating monster for map %s.", f.Id())
